move some rig stuff we may keep
