+++ pacman/operations/rig_algorithms/links.py
